refactor: split some test for performance
